servers/requirements.txt
alabaster==0.7.13
algokit==1.3.0
algokit-client-generator==1.0.2
algokit-utils==1.3.1
anaconda==0.0.1.1
anyio==3.7.1
argcomplete==3.0.8
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
Arpeggio==2.0.2
arrow==1.2.3
astroid==2.15.5
asttokens==2.2.1
astunparse==1.6.3
attrs==22.2.0
Babel==2.12.1
backcall==0.2.0
bardapi==0.1.38
bcrypt==4.0.1
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
browser-cookie3==0.19.1
btrfsutil==6.3.3
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
chardet==5.2.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
configobj==5.0.8
contourpy==1.1.0
copier==7.2.0
cryptography==41.0.3
cycler==0.11.0
dbus-python==1.3.2
debugpy==1.6.7
decorator==5.1.1
deep-translator==1.11.4
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.6
distro==1.8.0
docker==6.1.3
dockerpty==0.4.1
docopt==0.6.2
docutils==0.20.1
dunamai==1.16.1
executing==1.2.0
fastapi==0.103.1
fastjsonschema==2.17.1
fica==0.2.2
Flask==2.3.2
fonttools==4.41.1
fqdn==1.5.1
funcy==2.0
google-api-core==2.12.0
google-auth==2.23.2
google-cloud-core==2.3.3
google-cloud-translate==3.12.0
googleapis-common-protos==1.60.0
grpcio==1.59.0
grpcio-status==1.59.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httptools==0.6.0
httpx==0.23.3
hyperframe==6.0.1
i==2.1.0
idna==3.4
imagesize==1.4.1
ipykernel==6.23.2
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
isort==5.12.0
itsdangerous==2.1.2
jedi==0.18.2
jeepney==0.8.0
Jinja2==3.1.2
jinja2-ansible-filters==1.3.2
jsonpointer==2.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.1
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
jupytext==1.14.7
kiwisolver==1.4.4
langdetect==1.0.9
lazy-object-proxy==1.9.0
libtorrent==2.0.9
lz4==4.3.2
Mako==1.2.4
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mdit-py-plugins==0.4.0
mdurl==0.1.2
meson==1.2.1
mistune==2.0.5
mpmath==1.3.0
msgpack==1.0.5
mslex==1.1.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.5.0
nbformat==5.9.0
nest-asyncio==1.5.6
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.25.1
otter-grader==5.0.1
overrides==7.3.1
packaging==23.1
pandas==2.0.3
pandoc==2.3
pandocfilters==1.5.0
paramiko==3.2.0
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pipx==1.2.0
platformdirs==3.5.3
plumbum==1.8.1
ply==3.11
prometheus-client==0.17.0
prompt-toolkit==3.0.38
proto-plus==1.22.3
protobuf==4.24.4
psutil==5.9.5
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
py-algorand-sdk==2.3.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.24.0
pyclip==0.7.0
pycparser==2.21
pycryptodomex==3.18.0
pydantic==1.10.7
Pygments==2.15.1
PyGObject==3.44.1
pylint==2.17.4
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-on-whales==0.63.0
pytz==2023.3
pywal==3.3.0
PyYAML==5.4.1
pyyaml-include==1.3
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
questionary==1.10.0
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rsa==4.9
scipy==1.11.1
Send2Trash==1.8.2
shellingham==1.5.0.post1
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.4.1
Sphinx==7.1.0
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
tealish==0.0.2
terminado==0.17.1
terminator==2.1.3
texttable==1.6.7
textX==3.1.1
tinycss2==1.2.1
toml==0.10.2
tomlkit==0.11.8
tornado==6.3.2
tqdm==4.66.1
traitlets==5.9.0
typer==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
uri-template==1.2.0
urllib3==1.26.15
userpath==1.8.0
uvicorn==0.23.2
uvloop==0.17.0
validate==5.0.8
watchfiles==0.20.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==0.59.0
websockets==11.0.3
Werkzeug==2.3.6
widgetsnbextension==4.0.7
wrapt==1.15.0
